server: Refactor commands code, #125
